refactor: initializeTrashType 변수명 명시
